check file exists before trying to source cmp

+++ artisan.plugin.zsh
@@ -114,7 +114,8 @@ function _artisan() {
       # Cache invalidation check
       if [[ ! -f "$cache_file" || "$artisan_path" -nt "$cache_file" ||
         (-f "$composer_lock" && "$composer_lock" -nt "$cache_file") ||
-        ! cmp -s "$current_command_list" "$cache_file" ]]; then
+        (-f "$current_command_list" && ! cmp -s "$current_command_list" "$cache_file") ]]; then
+
         mv "$current_command_list" "$cache_file"
       else
         rm "$current_command_list"
